Restructure files again, add jest testing
