+++ package.json
@@ -44,6 +44,7 @@
     "@types/express": "^4.17.12",
     "@types/helmet": "^4.0.0",
     "@types/http-status": "^1.1.2",
+    "@types/jest": "^26.0.23",
     "@types/method-override": "0.0.31",
     "@types/node": "^15.6.1",
     "@types/supertest": "^2.0.11",
